:recycle: (PC-5679) refacto, move to a component
