+++ src/features/offer/components/index.ts
@@ -2,3 +2,4 @@ export { AccordionItem } from './AccordionItem'
 export { OfferHeader } from './OfferHeader'
 export { OfferHero } from './OfferHero'
 export { OfferIconCaptions } from './OfferIconCaptions'
+export { OfferWhereSection } from './OfferWhereSection'
