+++ src/features/offer/components/useDistance.ts
@@ -2,8 +2,8 @@ import { useGeolocation } from 'libs/geolocation'
 import { formatDistance } from 'libs/parsers'
 
 export const useDistance = (offerPosition: {
-  lat: number | undefined
-  lng: number | undefined
+  lat?: number
+  lng?: number
 }): string | undefined | null => {
   const userPosition = useGeolocation()
   if (!userPosition || !offerPosition) return null
